+++ tests/testthat/test-addin.R
@@ -1,43 +1,14 @@
 context("use_addin")
 
-test_that("use_addin() creates the first addins.dcf as promised, default", {
-  scoped_temporary_package()
-  use_addin()
-
-  addin_dcf <- readLines(proj_path("inst", "rstudio", "addins.dcf"))
-  expect_true(
-    all(
-      c("Name: New Addin Name", "Description: New Addin Description",
-        "Binding: new_addin", "Interactive: false") %in% addin_dcf
-    )
-  )
-})
-
-test_that("use_addin() creates the first addins.dcf as promised, option", {
+test_that("use_addin() creates the first addins.dcf as promised", {
   scoped_temporary_package()
   use_addin("addin.test")
 
   addin_dcf <- readLines(proj_path("inst", "rstudio", "addins.dcf"))
-  expect_true(
-    all(
-      c("Name: New Addin Name", "Description: New Addin Description",
-        "Binding: addin.test", "Interactive: false") %in% addin_dcf
-    )
-  )
+  expected_file <- system.file("templates", "addins.dcf", package = "usethis")
+  addin_dcf_expected <- readLines(expected_file)
+  addin_dcf_expected[3] <- "Binding: addin.test"
+  addin_dcf_expected[5] <- ""
+  expect_equal(addin_dcf, addin_dcf_expected)
 })
 
-test_that("use_addin() can create bindings for multiple addins.", {
-  scoped_temporary_package()
-  use_addin("addin1")
-  use_addin("addin_2")
-
-  addin_dcf <- readLines(proj_path("inst", "rstudio", "addins.dcf"))
-  expect_true(
-    all(
-      c("Name: New Addin Name", "Description: New Addin Description",
-        "Binding: addin1", "Interactive: false", "",
-        "Name: New Addin Name", "Description: New Addin Description",
-        "Binding: addin_2", "Interactive: false", "") %in% addin_dcf
-    )
-  )
-})