+++ NEWS.md
@@ -48,6 +48,9 @@
 
 * `use_release_issue()` creates a GitHub issue containing a release checklist 
   capturing best practices discovered by the tidyverse team (#338)
+  
+* New `use_addin()` helps setup necessary binding information for RStudio 
+  addins. (#591, #353)
 
 ## Partial file management